requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
praw==7.7.1
snscrape==0.7.0.20230622

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.5
xgboost==2.0.2

# NLP
textblob==0.17.1
spacy==3.7.2
wordcloud==1.9.3

# Visualization
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1
joblib==1.3.2

# Development
jupyter==1.0.0
ipykernel==6.27.1
black==23.12.0
pylint==3.0.3
pytest==7.4.3
